[kubernetes] Simplify sourcing of config-map ENV
Prior to kubernetes 1.6 (release planned april 2017) it is not
possible to cleanly import a config map containing hundreds of
key/value pairs as ENV var inside a pod.  One must specify each
variable individually which is not helpful in our case.

At this point the best we can do is to mount this config map within
the filesytem of the pods at run time and have a bashrc script
load those key/value pair for us.

Once Kubernetes 1.6 is out we will be able to revisit that part.